+++ src/pysm3/utils/__init__.py
@@ -16,8 +16,20 @@ from .photometry import (
     healpix_aperture_photometry,  # noqa: F401
 )
 from .small_scales import sigmoid  # noqa: F401
+from .spherical_harmonics import apply_smoothing_and_coord_transform, map2alm
 
 log = logging.getLogger("pysm3")
+
+
+def set_verbosity(level=logging.INFO):
+    logger = logging.getLogger("pysm3")
+    logger.setLevel(level)
+    if not logger.hasHandlers():
+        handler = logging.StreamHandler()
+        handler.setFormatter(
+            logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+        )
+        logger.addHandler(handler)
 
 
 def get_relevant_frequencies(freqs, low, high):
@@ -153,9 +165,7 @@ def bandpass_unit_conversion(
         weights /= np.trapz(weights, freqs)
         if weights.min() < cut:
             good = np.logical_not(weights < cut)
-            log.info(
-                f"Removing {good.sum()}/{len(good)} points below {cut}"
-            )
+            log.info(f"Removing {(good==0).sum()}/{len(good)} points below {cut}")
             weights = weights[good]
             freqs = freqs[good]
             weights /= np.trapz(weights, freqs)