+++ tests/test_smoothing.py
@@ -13,8 +13,9 @@ import pytest
 from astropy.tests.helper import assert_quantity_allclose
 
 import pysm3.units as u
-from pysm3.models import apply_smoothing_and_coord_transform
+from pysm3 import apply_smoothing_and_coord_transform
 
+INITIAL_FWHM = (1 * u.deg).to_value(u.radian)
 FWHM = (5 * u.deg).to_value(u.radian)
 NSIDE = 128
 CAR_RESOL = 12 * u.arcmin
@@ -25,15 +26,30 @@ LMAX = int(NSIDE * 1.5)
     scope="module"
 )  # scope makes the fixture just run once per execution of module
 def input_map():
-    beam_window = hp.gauss_beam(fwhm=FWHM, lmax=LMAX) ** 2
+    beam_window = hp.gauss_beam(fwhm=INITIAL_FWHM, lmax=LMAX) ** 2
     cl = np.zeros((6, len(beam_window)))
     cl[0:3] = beam_window
     np.random.seed(7)
     return hp.synfast(cl, NSIDE, lmax=LMAX, new=True) * u.uK_RJ
 
 
-def test_smoothing_healpix(input_map):
+def test_smoothing_healpix_beamwindow(input_map):
+    beam_window = hp.gauss_beam(fwhm=FWHM, lmax=LMAX, pol=True)
 
+    smoothed_map = apply_smoothing_and_coord_transform(
+        input_map, lmax=LMAX, beam_window=beam_window
+    )
+    assert input_map.shape[0] == 3
+    assert smoothed_map.shape == input_map.shape
+    assert_quantity_allclose(
+        actual=smoothed_map,
+        desired=hp.smoothing(input_map, fwhm=FWHM, lmax=LMAX, use_pixel_weights=True)
+        * input_map.unit,
+        rtol=1e-7,
+    )
+
+
+def test_smoothing_healpix(input_map):
     smoothed_map = apply_smoothing_and_coord_transform(
         input_map, lmax=LMAX, fwhm=FWHM * u.radian
     )
@@ -47,7 +63,6 @@ def test_smoothing_healpix(input_map):
 
 
 def test_car_nosmoothing(input_map):
-
     # `enmap_from_healpix` has no iteration or weights
     # so for test purpose we reproduce it here
     alm = (
@@ -71,13 +86,11 @@ def test_car_nosmoothing(input_map):
         pixell.reproject.enmap_from_healpix(
             input_map, shape, wcs, lmax=LMAX, rot=None, ncomp=3
         )
-        * input_map.unit
     )
     assert_quantity_allclose(actual=car_map, desired=map_rep)
 
 
 def test_healpix_output_nside(input_map):
-
     output_nside = 64
     output_map = apply_smoothing_and_coord_transform(
         input_map, fwhm=None, output_nside=output_nside, lmax=LMAX